soc: mediatek: Bind missing clocks to scpsys on the MT6797
Add missing clocks to MJC, AUDIO, VENC, VDEC domains. Also add MM clock
to VENC and VDEC domains, since the power domains are initially turned
on in the order of the scp_domain_data array, not in dependency order.
VENC and VDEC are subdomains of MM, but MM is defined after them, so
they are turned on before their parent. If the MM domain is not
initially on, boot will fail. This is the case e.g. with kexec boot.

Signed-off-by: Jasper Mattsson <jasu@njomotys.info>

addclk