soc: mediatek: Allow gaps in SCP PDs
The power domains array for MT6797 contains gaps, since most MFG-related
power domains are missing. This causes a lookup for a regulator with
a NULL name fail, and this will cause the probe of SCPSYS to fail. This
patch skips gaps in the array by detecting if the power domain name is
NULL. This could all be easily solved by reordering the #defines in
dt-bindings, but that could break existing DTBs.

Signed-off-by: Jasper Mattsson <jasu@njomotys.info>